@@ -5,7 +5,7 @@ author: PlatiB <platib85@gmail.com>
 homepage: https://github.com/PlatiB/korean_words
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
